+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.5.0-beta
 
 - Adds `PurgeInstancesMetadataAsync` and `PurgeInstancesAsync` support and implementation to `DurableTaskGrpcClient`
+- Fix issue with mixed Newtonsoft.Json and System.Text.Json serialization.
 
 ### Breaking changes